move 3 files about Character Feature to a new folder
